Refine transactions feature: integer amounts, validation, and display formatting
- Change debitAmount/creditAmount from Double to Int (store cents) with DB migration 2→3
- Add TransactionValidator for double-entry balance checks
- Add TransactionDisplay for formatting amounts and entry labels
- Update AddTransaction and TransactionDetail UIs with improved layout
- Add unit tests for validator and display formatting

+++ app/src/main/java/dev/fitiavana/accounting/data/model/TransactionEntry.kt
@@ -30,6 +30,6 @@ data class TransactionEntry(
     @PrimaryKey val id: String,
     val transactionId: String,
     val accountId: String,
-    val debitAmount: Double?,
-    val creditAmount: Double?
+    val debitAmount: Int?,
+    val creditAmount: Int?
 )
